Integrate samples totalvi fix (#37)
* fixing a bug with totalvi model file

* adding the integration model to file names of figures

* add flag to allow not removing working dir from tissue_integration_results run via dashboard_utils.py

* fixing rm commands

+++ data_processing/scripts/dashboard_utils.py
@@ -2,7 +2,8 @@
 # - To get tissue coverage csv, run as follows:
 #   python dashboard_utils.py tissue_coverage
 # - To get tissue integration results csv, run as follows:
-#   python dashboard_utils.py tissue_integration_results <working_dir> <s3_access_file>
+#   python dashboard_utils.py tissue_integration_results <working_dir> <s3_access_file> <rm_working_dir>
+#   where <rm_working_dir> is "True" or "False" for indicating whether the working dir should be removed at the end; this argument is optional (default value is "True").
 
 import sys
 import io
@@ -100,29 +101,31 @@ def generate_tissue_integration_figures(adata, tissue, version, working_dir, bas
     # use umap embeddings based on totalVI if available; otherwise based on scvi
     if "X_umap_totalvi_integrated" in adata.obsm:
         umap_key = "X_umap_totalvi_integrated"
+        integration_model = "totalvi"
     else:
         umap_key = "X_umap_scvi_integrated"
+        integration_model = "scvi"
     adata.obsm["X_umap"] = adata.obsm[umap_key]
     
     for field in fields_to_plot:
         # looks like sc.pl.umap cannot take an absolute path; move files from the output dir of sc.pl.umap to the working directory
-        filename = ".{}.{}.pdf".format(prefix,field)
+        filename = ".{}.{}.{}.pdf".format(prefix,integration_model,field)
         sc.pl.umap(adata, color=[field], save = filename)
         filename = "umap" + filename
-        os.rename("figures/" + filename, os.path.join(figures_dir, filename))
+        shutil.move("figures/" + filename, os.path.join(figures_dir, filename))
 
     # for each celltypist model used plot a umap colored by the predicted cell types
     mdl = 1
     while "celltypist_model.{}".format(str(mdl)) in adata.obs:
         mdl_name = "celltypist_{}".format(adata.obs["celltypist_model." + str(mdl)][1].split("/")[-1].split(".")[0])
-        filename = ".{}.{}.pdf".format(prefix,mdl_name)
+        filename = ".{}.{}.{}.pdf".format(prefix,integration_model,mdl_name)
         sc.pl.umap(adata, color=["celltypist_predicted_labels.{}".format(str(mdl))], save = filename)
         filename = "umap" + filename
-        os.rename("figures/" + filename, os.path.join(figures_dir, filename))
+        shutil.move("figures/" + filename, os.path.join(figures_dir, filename))
         mdl += 1
 
     logger.add_to_log("Saving .zip file with the figures...")
-    zip_filename = "{}.figures.zip".format(prefix)
+    zip_filename = "{}.{}.figures.zip".format(prefix,integration_model)
     figures_file_path = os.path.join(working_dir, zip_filename)
     zipf = zipfile.ZipFile(figures_file_path, 'w', zipfile.ZIP_DEFLATED)
     utils.zipdir(figures_dir, zipf)
@@ -135,10 +138,11 @@ def generate_tissue_integration_figures(adata, tissue, version, working_dir, bas
     logger.add_to_log("aws response: {}\n".format(os.popen(sync_cmd).read()))
     
     shutil.rmtree("figures")
+    shutil.rmtree(figures_dir)
     figures_url = "{}{}/{}/{}/{}".format(base_aws_url, base_s3_dir, tissue, version, zip_filename)
     return figures_url
 
-def get_tissue_integration_results_csv(working_dir: str, s3_access_file: str):
+def get_tissue_integration_results_csv(working_dir: str, s3_access_file: str, rm_working_dir: bool):
     logger.add_to_log("Downloading the Samples sheet from Google drive...")
     samples = utils.read_immune_aging_sheet("Samples")
     tissues = np.unique(samples["Organ"])
@@ -217,11 +221,13 @@ def get_tissue_integration_results_csv(working_dir: str, s3_access_file: str):
                 # generate figures
                 csv_row[CSV_HEADER_FIGURES] = generate_tissue_integration_figures(adata, tissue, version, working_dir, BASE_AWS_URL, BASE_S3_URL, BASE_S3_DIR)
 
-                # clean up
-                os.system("rm -r {}/*".format(working_dir))
+                if rm_working_dir:
+                    # clean up
+                    os.system("rm -r {}/*".format(working_dir))
         except Exception:
             logger.add_to_log("Execution failed for tissue {} with the following error:\n{}".format(tissue, traceback.format_exc()), "critical")
-            os.system("rm -r {}/*".format(working_dir))
+            if rm_working_dir:
+                os.system("rm -r {}/*".format(working_dir))
             logger.add_to_log("Continuing execution for other tissues...")
             continue
         csv_rows.append(csv_row)
@@ -255,4 +261,8 @@ else:
     s3_access_file = sys.argv[3] # must be the absolute path to the aws credentials file
     assert(os.path.isdir(working_dir))
     assert(os.path.isfile(s3_access_file))
-    get_tissue_integration_results_csv(working_dir, s3_access_file)
+    rm_working_dir = True
+    if len(sys.argv) > 4:
+        assert(sys.argv[4] == "True" or sys.argv[4] == "False")
+        rm_working_dir = sys.argv[4] == "True"
+    get_tissue_integration_results_csv(working_dir, s3_access_file, rm_working_dir)
